Refactor to own folder
